# frontend/vue-project/merged-api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://10.239.105.110:11082",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/recycle/restore": {
      "post": {
        "tags": [
          "回收站管理"
        ],
        "summary": "还原文件/文件夹",
        "description": "还原文件或文件夹",
        "operationId": "restore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/rename": {
      "post": {
        "tags": [
          "文件管理"
        ],
        "summary": "重命名文件/文件夹",
        "description": "重命名一个文件或文件夹",
        "operationId": "rename",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/move-to": {
      "post": {
        "tags": [
          "文件管理"
        ],
        "summary": "移动文件/文件夹",
        "description": "将文件/文件夹移动到指定父目录下",
        "operationId": "moveTo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/folders": {
      "post": {
        "tags": [
          "文件管理"
        ],
        "summary": "新建文件夹",
        "description": "在指定的父目录下创建一个新文件夹",
        "operationId": "createFolder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFolderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recycle": {
      "get": {
        "tags": [
          "回收站管理"
        ],
        "summary": "获取回收站文件列表",
        "description": "查询回收站内所有文件和文件夹",
        "operationId": "listRecycleBinFiles",
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "description": "页码",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1001
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "页大小",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 50
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageResultRecycleFileVO"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/list-folderOnly": {
      "get": {
        "tags": [
          "文件管理"
        ],
        "summary": "仅获取文件夹列表",
        "description": "根据父目录ID查询该目录下的子文件夹",
        "operationId": "listFolderOnly",
        "parameters": [
          {
            "name": "parentId",
            "in": "query",
            "description": "父目录ID，若为0或null则自动定位到用户根目录",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1001
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListVirtualFileVO"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/list-file": {
      "get": {
        "tags": [
          "文件管理"
        ],
        "summary": "获取文件列表",
        "description": "根据父目录ID查询该目录下的所有文件和子文件夹",
        "operationId": "listFiles",
        "parameters": [
          {
            "name": "parentId",
            "in": "query",
            "description": "父目录ID，若为0或null则自动定位到用户根目录",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1001
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListVirtualFileVO"
                }
              }
            }
          }
        }
      }
    },
    "/api/recycle/permanent": {
      "delete": {
        "tags": [
          "回收站管理"
        ],
        "summary": "批量彻底删除",
        "description": "彻底删除文件/文件夹",
        "operationId": "deletePermanently",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/files": {
      "delete": {
        "tags": [
          "文件管理"
        ],
        "summary": "批量移入回收站",
        "description": "逻辑删除，将文件/文件夹移入回收站",
        "operationId": "moveToRecycleBin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "认证模块"
        ],
        "summary": "注册",
        "description": "根据登录凭证（目前为用户名）和密码注册",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultRegisterResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "认证模块"
        ],
        "summary": "登录",
        "description": "根据登录凭证（目前为用户名）和密码登录",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLoginResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "description": "通用响应对象"
      },
...
      "FileItemType": {
        "type": "string",
        "description": "文件/文件夹类型",
        "example": "FILE",
        "enum": [
          "FILE",
          "FOLDER"
        ]
      },
      "ItemIdentity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "文件/文件夹ID",
            "format": "int64",
            "example": 1001
          },
          "type": {
            "$ref": "#/components/schemas/FileItemType"
          }
        },
        "description": "简易对象实体",
        "example": [
          {
            "id": 100,
            "type": "FILE"
          },
          {
            "id": 200,
            "type": "FOLDER"
          }
        ]
      },
      "RestoreRequest": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "待还原的文件/文件夹标识列表",
            "example": [
              {
                "id": 100,
                "type": "FILE"
              },
              {
                "id": 200,
                "type": "FOLDER"
              }
            ],
            "items": {
              "$ref": "#/components/schemas/ItemIdentity"
            }
          }
        },
        "description": "批量还原请求体"
      },
      "RenameRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "文件/文件夹Id",
            "format": "int64",
            "example": 1001
          },
          "newName": {
            "type": "string",
            "description": "新名字",
            "example": "new Name"
          },
          "type": {
            "$ref": "#/components/schemas/FileItemType"
          }
        },
        "description": "用于文件/文件夹重命名的请求体"
      },
      "MoveRequest": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "待删除的文件/文件夹标识列表（支持混合删除）",
            "example": [
              {
                "id": 100,
                "type": "FILE"
              },
              {
                "id": 200,
                "type": "FOLDER"
              }
            ],
            "items": {
              "$ref": "#/components/schemas/ItemIdentity"
            }
          },
          "parentId": {
            "type": "integer",
            "description": "目标父目录Id",
            "format": "int64",
            "example": 1001
          }
        },
        "description": "移动文件/文件夹的请求体"
      },
      "CreateFolderRequest": {
        "required": [
          "name",
          "parentId"
        ],
        "type": "object",
        "properties": {
          "parentId": {
            "type": "integer",
            "description": "父目录ID，若为0表示根目录",
            "format": "int64",
            "example": 1001
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string",
            "description": "文件夹名称",
            "example": "工作文档"
          }
        },
        "description": "创建文件夹请求体"
      },
...
      "PageResultRecycleFileVO": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecycleFileVO"
            }
          }
        },
        "description": "响应数据"
      },
      "RecycleFileVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "文件/文件夹ID",
            "format": "int64",
            "example": 1001
          },
          "name": {
            "type": "string",
            "description": "文件/文件夹名称",
            "example": "name"
          },
          "size": {
            "type": "integer",
            "description": "文件大小（字节）",
            "format": "int64",
            "example": 1024
          },
          "deletedAt": {
            "type": "string",
            "description": "删除时间",
            "format": "date-time"
          },
          "type": {
            "$ref": "#/components/schemas/FileItemType"
          }
        },
        "description": "用于前端回收站展示的视图对象"
      },
      "ResultPageResultRecycleFileVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultRecycleFileVO"
          }
        },
        "description": "通用响应对象"
      },
      "ResultListVirtualFileVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "array",
            "description": "响应数据",
            "items": {
              "$ref": "#/components/schemas/VirtualFileVO"
            }
          }
        },
        "description": "通用响应对象"
      },
      "VirtualFileVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "文件/文件夹ID",
            "format": "int64",
            "example": 1001
          },
          "name": {
            "type": "string",
            "description": "文件/文件夹名称",
            "example": "name"
          },
          "size": {
            "type": "integer",
            "description": "文件大小（字节）",
            "format": "int64",
            "example": 1024
          },
          "updateTime": {
            "type": "string",
            "description": "最后更新时间",
            "format": "date-time"
          },
          "type": {
            "$ref": "#/components/schemas/FileItemType"
          },
          "parentId": {
            "type": "integer",
            "description": "父目录ID",
            "format": "int64",
            "example": 1001
          }
        },
        "description": "用于前端展示的视图对象"
      },
      "DeleteRequest": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "待删除的文件/文件夹标识列表（支持混合删除）",
            "example": [
              {
                "id": 100,
                "type": "FILE"
              },
              {
                "id": 200,
                "type": "FOLDER"
              }
            ],
            "items": {
              "$ref": "#/components/schemas/ItemIdentity"
            }
          }
        },
        "description": "批量删除请求体"
      },
...
      "RegisterRequest": {
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "maxLength": 20,
            "minLength": 3,
            "type": "string",
            "description": "用户名",
            "example": "aaabbbccc"
          },
          "email": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "邮箱",
            "example": "example@ex.com"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "description": "手机号码",
            "example": "+00 123456789"
          },
          "password": {
            "maxLength": 20,
            "minLength": 6,
            "type": "string",
            "description": "密码",
            "example": "123abc"
          }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "description": "响应数据"
      },
      "ResultRegisterResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/RegisterResponse"
          }
        },
        "description": "通用响应对象"
      },
      "LoginRequest": {
        "required": [
          "loginAccount",
          "password"
        ],
        "type": "object",
        "properties": {
          "loginAccount": {
            "type": "string",
            "description": "登录凭证（目前为用户名）",
            "example": "aaabbbccc"
          },
          "password": {
            "type": "string",
            "description": "登录密码",
            "example": "123abc"
          }
        },
        "description": "取消上传请求体"
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "登录token",
            "example": "aaabbbccc"
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "example": "nickname"
          },
          "username": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "description": "过期时间",
            "format": "int64"
          },
          "rootFolderId": {
            "type": "integer",
            "description": "根目录Id",
            "format": "int64",
            "example": 1001
          },
          "rootFolderName": {
            "type": "string",
            "description": "根目录名称",
            "example": "folderName"
          }
        },
        "description": "响应数据"
      },
      "ResultLoginResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/LoginResponse"
          }
        },
        "description": "通用响应对象"
      }
    }
  }
}